refactor: merge positive and negative validation tests into a single file test_new_item2 and remove @pytest.mark.skip from test_create_new_item
