+++ tests/test_create_new_item.py
@@ -4,8 +4,6 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 
 
-@pytest.mark.skip(reason="Test fails because Jenkins allows '|' character - temporary skip")
-@pytest.mark.skip(reason="fails in CI")
 @pytest.mark.parametrize("invalid_char", ["?", "*", "/", "|", "!", "%", "&", ";", ":"])
 def test_create_new_item_validate_unsupported_special_characters(browser, invalid_char):
     wait = WebDriverWait(browser, 10)
